+++ nemo/utils/__init__.py
@@ -23,4 +23,3 @@ logging = _Logger()
 from .argparse import NemoArgParser
 from .exp_logging import ExpManager, get_logger
 from .helpers import *
-from nemo.utils.app_state import AppState
