+++ nemo/utils/app_state.py
@@ -20,8 +20,8 @@
 # nothing from nemo.core, including e.g. types (so we cannot use them for "python 3 type hints").
 import nemo
 from nemo.utils.metaclasses import Singleton
-from nemo.utils.neural_graph_manager import NeuralGraphManager
-from nemo.utils.object_registry import ObjectRegistry
+from nemo.utils.neural_graph.neural_graph_manager import NeuralGraphManager
+from nemo.utils.neural_graph.object_registry import ObjectRegistry
 
 
 class AppState(metaclass=Singleton):
