+++ nemo/utils/neural_graph/connection.py
